third_party/blink/web_tests/dom/mutation-event-tests/fast/forms/select/select-set-length-with-mutation-reorder-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Tests that setting the .length on an HTMLSelectElement works in the presence of DOM mutation listeners that reorder option elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.length is 4
PASS sel.length is 2
PASS sel.options.item(0).id is "a"
PASS sel.options.item(1).id is "b"
PASS successfullyParsed is true

TEST COMPLETE

a
b
